requirements.txt
django
djangorestframework 
django-cors-headers
djangorestframework-simplejwt
sendgrid==6.12.5
python-http-client==3.3.7 
python-dotenv
gunicorn
whitenoise
dj-database-url
psycopg2-binary